+++ docs/changes.rst
@@ -31,6 +31,7 @@ Pending
   needing to respect ``django.conf.settings.DEBUG`` to match
   ``debug_toolbar_urls``.
 * Fixed cookie ``expires`` calculation in ``djdt.cookie.set``.
+* Account for the new ``CULL_PROBABILITY`` in Django 6.2 in tests.
 
 6.3.0 (2026-04-01)
 ------------------